Image Encryption and Decryption using DES
This is a mini project wherein I have tried to encode an image in .jpg format into a byte stream array and then to Unicode characters and finally tried to decode it and get back the original image in .jpg format. All this has been done using the encryption technique in cryptography called DES (Data Encryption Standard) which is a symmetric key encryption which means that it uses the same key for both encryptions as well as decryptions.