Add files via upload
Encimg.jpg is the input file given for encryption
Imgoutput.jpg is the file obtained after decryption
EncJpgImgOut.txt is the output text file containing encoded characters in Unicode.
DecJpgImgOut.txt is the Output text file containing decoded characters in Unicode.
DemoData.txt file contains a sample input sentence to check whether encoding and decoding are properly done or not.
encypteddata.txt is the file containing the characters of the DemoData.txt file in an encoded format.
decrypteddata.txt file contains the decoded characters of the encrypteddata.txt file. If this file returns the original text contents of DemoData.txt file, then we can confirm that the encoding and decoding process has been done properly.